+++ vitis/software/Baremetal/test/uz/uz_dqn/test_uz_bitflip_V2.c
@@ -13,6 +13,7 @@
 #include <time.h>
 #include <string.h>
 #include <stdlib.h>
+#include "export_array.h"
 
 void uz_nn_trained_export(uz_nn_t *self);
 
@@ -275,8 +276,8 @@ void test_dqn_bitflip(void)
     if (f != NULL)                       // check for success
     {
      fprintf(f,"Learnrate, Discount Factor,Epsilon_start,Epsilon_min,Epsilon_decay,Hidden Layer,Bufferlength,Minibatchsize,Epochen,Targetupdatefrequency,Numberofbits,Numberofneuronsinhiddenlayer \n");
-     fprintf(f,"%.6f,%.6f,%.6f,%.6f,%.6f,%d,%d,%d,%d,%d,%d,%d\n", lernrate,discountfact, configenv.epsilon_start,configenv.epsilon_min,configenv.epsilon_decay,NUMBER_OF_HIDDEN_LAYER,EXPERIENCE_BUFFER_LENGTH,MINIBATCHSIZE,NUMBER_OF_EPOCHS,TARGET_UPDATE_FREQUENCY,NUMBEROFBITS,
-     NUMBER_OF_NEURONS_IN_HIDDEN_LAYER);
+     fprintf(f, "%.6f,%.6f,%.6f,%.6f,%.6f,%d,%d,%d,%d,%d,%d,%d\n", (double)lernrate, (double)discountfact, (double)configenv.epsilon_start, (double)configenv.epsilon_min, (double)configenv.epsilon_decay, NUMBER_OF_HIDDEN_LAYER, EXPERIENCE_BUFFER_LENGTH, MINIBATCHSIZE, NUMBER_OF_EPOCHS, TARGET_UPDATE_FREQUENCY, NUMBEROFBITS,
+             NUMBER_OF_NEURONS_IN_HIDDEN_LAYER);
      fclose(f);                       // close the file
      f = NULL;                        // set file handle to null since f is no longer valid
     }
